Use BigDecimal instead of long for the amount

+++ sentinel-generator/src/main/java/dev/delath/sentinel/events/TransactionEvent.java
@@ -2,12 +2,14 @@ package dev.delath.sentinel.events;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 public record TransactionEvent(
     @JsonProperty("transaction_id") String transactionId,
     @JsonProperty("user_id") String userId,
     @JsonProperty("card_token") String cardToken,
     @JsonProperty("card_bin") String cardBin,
-    long amount,
+    BigDecimal amount,
     String currency,
     @JsonProperty("merchant_details") MerchantDetails merchantDetails,
     Location location,
